requirements.txt
python-telegram-bot==13.7
pymongo==3.12.0
dnspython==2.1.0